requirements.txt
altgraph==0.17.4
click==8.2.1
macholib==1.16.3
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==25.0
pillow==11.2.1
Pygments==2.19.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
rich==14.0.0
setuptools==80.9.0
shellingham==1.5.4
typer==0.16.0
typing_extensions==4.14.0
-e git+ssh://git@github.com/andriawan24/webp-converter.git@e14a83c17e0dd31adb2b3629585b662164362513#egg=webp_converter
